chore: replace `useModalModelVersionDelete` with `ModelVersionDeleteModal`
